Rename `common.ts` -> 'thorchainCommon.ts`
Make it clear this isn't a general common file, but specific to Thorchain
